+++ admin/header.php
@@ -4,6 +4,9 @@ require_once dirname(__DIR__).'/init.php';
 
 // Include functions
 require_once PATH.ADMIN.INC.'/functions.php';
+
+// Fetch the current page
+$current_page = getCurrentPage();
 ?>
 <!DOCTYPE html>
 <html>
@@ -16,22 +19,22 @@ require_once PATH.ADMIN.INC.'/functions.php';
 		<?php getStylesheet('buttons.css'); ?>
 		<?php getAdminStylesheet('style.css'); ?>
 	</head>
-	<body>
+	<body class="<?php echo $current_page; ?>">
 		<header id="admin-header">
 			<a id="site-title" href="/"><?php getSetting('site_title'); ?></a>
 		</header>
 		<div id="admin-nav-wrap"></div>
-		<nav id="admin-nav">
-			<ul class="nav">
+		<nav id="admin-nav-menu">
+			<ul class="menu">
 				<?php
-				adminNavItem('Dashboard', 'index.php');
-				adminNavItem('Pages', '', array(array('List Pages', 'Create Page'), array('posts.php?type=page', 'posts.php?type=page&action=create')));
-				adminNavItem('Posts', '', array(array('List Posts', 'Create Post', 'List Categories'), array('posts.php', 'posts.php?action=create', 'categories.php')));
-				adminNavItem('Media', '', array(array('List Media', 'Upload Media'), array()));
-				adminNavItem('Navigation', '', array(array('', ''), array()));
-				adminNavItem('Widgets', '', array(array('List Widgets', 'Create Widget'), array()));
-				adminNavItem('Users', '', array(array('List Users', 'Create User', 'Your Profile'), array('users.php', 'users.php?action=create', '')));
-				adminNavItem('Settings', 'settings.php');
+				adminNavMenuItem('dashboard', 'index.php');
+				adminNavMenuItem('pages', '', array(array('List Pages', 'Create Page'), array('posts.php?type=page', 'posts.php?type=page&action=create')));
+				adminNavMenuItem('posts', '', array(array('List Posts', 'Create Post', 'List Categories'), array('posts.php', 'posts.php?action=create', 'categories.php')));
+				adminNavMenuItem('media', '', array(array('List Media', 'Upload Media'), array()));
+				adminNavMenuItem('navigation', '', array(array('', ''), array()));
+				adminNavMenuItem('widgets', '', array(array('List Widgets', 'Create Widget'), array()));
+				adminNavMenuItem('users', '', array(array('List Users', 'Create User', 'Your Profile'), array('users.php', 'users.php?action=create', '')));
+				adminNavMenuItem('settings', 'settings.php');
 				?>
 			</ul>
 		</nav>